Added routed to lib folder.
